Restructure into documentdb_tests/ package with core/ test hierarchy
Move all source files under documentdb_tests/ top-level package and
reorganize tests into the core/ folder structure following the MongoDB
documentation hierarchy.

Structure:
  documentdb_tests/
    conftest.py
    pytest.ini
    framework/              # Shared testing framework
    compatibility/
      result_analyzer/      # Test result analysis tools
      tests/
        core/
          collections/capped_collections/
          operator/stages/{group,match}/
          query-and-write/commands/{find,insert}/

File moves (no content changes beyond import paths):
- framework/ -> documentdb_tests/framework/
- result_analyzer/ -> documentdb_tests/compatibility/result_analyzer/
- tests/collection/ -> tests/core/collections/capped_collections/
- tests/aggregate/ -> tests/core/operator/stages/{group,match}/
- tests/find/ -> tests/core/query-and-write/commands/find/
- tests/insert/ -> tests/core/query-and-write/commands/insert/

Add docs/testing/FOLDER_STRUCTURE.md documenting the layout.
