+++ .gitignore
@@ -208,3 +208,4 @@ __marimo__/
 
 # Test results directory
 .test-results/
+.idea/
