+++ documentdb_tests/compatibility/tests/__init__.py
@@ -7,7 +7,7 @@ This package provides test utilities and common assertions for testing
 DocumentDB functionality.
 """
 
-from framework.assertions import (
+from documentdb_tests.framework.assertions import (
     assert_count,
     assert_document_match,
     assert_documents_match,
