+++ documentdb_tests/compatibility/tests/core/collections/capped_collections/test_capped_collections.py
