+++ documentdb_tests/compatibility/tests/core/operator/stages/group/test_group_stage.py
