+++ documentdb_tests/compatibility/tests/core/operator/stages/match/test_match_stage.py
