+++ documentdb_tests/compatibility/tests/core/query-and-write/commands/find/test_projections.py
@@ -6,7 +6,7 @@ Tests for field inclusion, exclusion, and projection operators.
 
 import pytest
 
-from framework.assertions import assert_field_exists, assert_field_not_exists
+from documentdb_tests.framework.assertions import assert_field_exists, assert_field_not_exists
 
 
 @pytest.mark.find
