+++ documentdb_tests/compatibility/tests/core/query-and-write/commands/find/test_query_operators.py
