+++ documentdb_tests/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 # Test discovery
-testpaths = tests
+testpaths = compatibility/tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
