--- tests/aggregate/__init__.py
@@ -1,1 +1,0 @@
-"""Aggregation pipeline tests."""
