--- tests/collection/__init__.py
@@ -1,1 +1,0 @@
-"""Collection management tests."""
