--- tests/find/__init__.py
@@ -1,1 +1,0 @@
-"""Find operation tests."""
