--- tests/insert/__init__.py
@@ -1,1 +1,0 @@
-"""Insert operation tests."""
